@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.3",
+        "php": "^7.4",
         "ext-gd":" *",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^1.0",
