@@ -3,18 +3,32 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasOne;
 
+/**
+ * @property int $id
+ * @property string $text
+ * @property integer $sender
+ * @property integer $receiver
+ * @property bool $read
+ * @property User $fromContact
+ */
 class Message extends Model
 {
-
   /**
    * Fields that are mass assignable
    *
    * @var array
    */
-  protected $fillable = ['text','sender', 'receiver', 'read'];
+  protected $fillable = [
+      'text',
+      'sender',
+      'receiver',
+      'read'
+  ];
 
-  public function user()
+  public function user(): BelongsTo
   {
     return $this->belongsTo(User::class);
   }
@@ -22,7 +36,7 @@ class Message extends Model
   /**
    * loads the user object from the message object
   */
-  public function fromContact()
+  public function fromContact(): HasOne
   {
     return $this->hasOne(User::class, 'id', 'sender');
   }
