@@ -2,6 +2,8 @@
 
 namespace App\Providers;
 
+use App\Events\NewFriendRequestEvent;
+use App\Listeners\NewFriendRequestListener;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
@@ -17,7 +19,9 @@ class EventServiceProvider extends ServiceProvider
         Registered::class => [
             SendEmailVerificationNotification::class,
         ],
-
+        NewFriendRequestEvent::class => [
+            NewFriendRequestListener::class,
+        ],
         'App\Events\Event' => [
            'App\Listeners\EventListener',
         ],
