@@ -1,37 +1,40 @@
-<?php 
+<?php
 
 namespace App\Traits;
+
 use App\User;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 trait Friend
 {
+
     /**
 	 * friendship that this user started
-	 * */ 
-	protected function friendsOfThisUser()
-	{
+	 * */
+	protected function friendsOfThisUser(): BelongsToMany
+    {
 		return $this->belongsToMany(User::class, 'friendships', 'first_user', 'second_user')
 		            ->withPivot('status')->wherePivot('status', 'confirmed');
 	}
- 
+
 	/**
 	 * friendship that this user was asked for
-	 * */ 
-	protected function thisUserFriendOf()
-	{
+	 * */
+	protected function thisUserFriendOf(): BelongsToMany
+    {
 		return $this->belongsToMany(User::class, 'friendships', 'second_user', 'first_user')
 		            ->withPivot('status')->wherePivot('status', 'confirmed');
 	}
- 
+
     /**
 	 * accessor allowing you call $user->friends
-	 * */ 
+	 * */
 	public function getFriendsAttribute()
 	{
 		if ( ! array_key_exists('friends', $this->relations)) $this->loadFriends();
 		    return $this->getRelation('friends');
 	}
- 
+
 	protected function loadFriends()
 	{
 		if ( ! array_key_exists('friends', $this->relations))
@@ -40,13 +43,13 @@ trait Friend
             $this->setRelation('friends', $friends);
         }
 	}
- 
+
 	protected function mergeFriends()
 	{
-		if($temp = $this->friendsOfThisUser)
-		    return $temp->merge($this->thisUserFriendOf);
-		else
-		    return $this->thisUserFriendOf;
+		if ($temp = $this->friendsOfThisUser){
+            return $temp->merge($this->thisUserFriendOf);
+        } else {
+            return $this->thisUserFriendOf;
+        }
 	}
-
 }
